fix: move rule button from start to home screen
